Update predicates.py after comments
Made changes to reflect comments by Chris:

1. Do not modify "ed50": f"RO:0002434" until we know where this predicate comes from. 
"But I guess the main point is that it makes no sense to have an ed50 between a chemical and a gene/protein. Is it possible to see whether we actually have any of these?"

2. "gi50": f"RO:0002434" -> {DGIDB}:Inhibitor

3. Do not modify "interacts_with"  until we know where this predicate comes from.
I" guess it would be easier to be sure if we know the context in which "interacts-with" was found in our ingestss.

(but I'd be tempted to put 2436 for this one)"

Main take away: where do these predicates come from? Can we track down how this was generated?